requirements.txt
asgiref==3.3.1
boto3==1.16.36
botocore==1.19.36
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
django-rest-knox==4.1.0
djangorestframework==3.12.2
dynaconf==3.1.2
hvac==0.10.5
idna==2.10
Jinja2==2.11.2
jmespath==0.10.0
knox==0.1.11
loguru==0.5.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
cryptography==2.8
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
validators==0.18.1
win32-setctime==1.0.3
